# internal/delivery/docs/swagger.yaml
definitions:
  handlers.payload:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  models.Metrics:
    properties:
      alltime_income:
        type: integer
      alltime_out:
        type: integer
      current_users:
        type: integer
      income_sub_out:
        type: integer
    type: object
  models.MetricsSet:
    properties:
      alltime_income:
        type: integer
      alltime_out:
        type: integer
      current_users:
        type: integer
    type: object
  models.Rate:
    properties:
      commission:
        type: integer
      id:
        type: integer
      min_lock_days:
        type: integer
      profit:
        type: integer
      properties: {}
      title:
        type: string
    type: object
  models.RateCreate:
    properties:
      commission:
        type: integer
      min_lock_days:
        type: integer
      profit:
        type: integer
      properties: {}
      title:
        type: string
    type: object
  models.Review:
    properties:
      id:
        type: integer
      properties: {}
      text:
        type: string
      tittle:
        type: string
    type: object
  models.ReviewCreate:
    properties:
      properties: {}
      text:
        type: string
      tittle:
        type: string
    type: object
  models.User:
    properties:
      email:
        type: string
      id:
        type: integer
      phone_number:
        type: string
      properties: {}
      refer_id:
        description: 'TODO: test маршалинг анмаршалинг и инсерт / select с бдшки'
        type: integer
      status:
        type: string
    type: object
  models.UserCreate:
    properties:
      email:
        type: string
      password:
        type: string
      phone_number:
        type: string
      properties: {}
      refer_id:
        description: 'TODO: test маршалинг анмаршалинг и инсерт / select с бдшки'
        type: integer
    type: object
  models.UserRate:
    properties:
      deposit:
        type: integer
      earned_pool:
        type: integer
      id:
        type: integer
      lock:
        type: string
      opened:
        type: string
      outcome_pool:
        type: integer
      rate_id:
        type: integer
      token:
        type: string
      user_id:
        type: integer
    type: object
  models.UserRateCreate:
    properties:
      deposit:
        type: integer
      lock:
        type: string
      rate_id:
        type: integer
      token:
        type: string
      user_id:
        type: integer
    type: object
info:
  contact: {}
paths:
  /admin/rate:
    post:
      consumes:
      - application/json
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.RateCreate'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully created
          schema:
            type: integer
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      tags:
      - admin
    put:
      consumes:
      - application/json
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Rate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      tags:
      - admin
  /admin/static/metrics:
    post:
      consumes:
      - application/json
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Metrics updated
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.MetricsSet'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated metrics
          schema:
            type: integer
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      tags:
      - admin
  /admin/static/review:
    delete:
      consumes:
      - application/json
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Review ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      tags:
      - admin
    post:
      consumes:
      - application/json
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Review create
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.ReviewCreate'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully created review
          schema:
            type: integer
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      tags:
      - admin
    put:
      consumes:
      - application/json
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Review
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Review'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      tags:
      - admin
  /admin/user/all:
    put:
      consumes:
      - application/json
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: page
        in: query
        name: page
        required: true
        type: integer
      - description: per page
        in: query
        name: per_page
        required: true
        type: integer
      - description: filter by status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      tags:
      - admin
  /public/rate:
    get:
      consumes:
      - application/json
      parameters:
      - description: Page
        in: query
        name: page
        required: true
        type: integer
      - description: Reviews per page
        in: query
        name: per_page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Array
          schema:
            items:
              $ref: '#/definitions/models.Rate'
            type: array
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      tags:
      - public
  /public/static/metrics:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: Successfully
          schema:
            $ref: '#/definitions/models.Metrics'
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      tags:
      - public
  /public/static/review:
    get:
      consumes:
      - application/json
      parameters:
      - description: Page
        in: query
        name: page
        required: true
        type: integer
      - description: Reviews per page
        in: query
        name: per_page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Array of reviews
          schema:
            items:
              $ref: '#/definitions/models.Review'
            type: array
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      tags:
      - public
  /public/user:
    post:
      consumes:
      - application/json
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.UserCreate'
      produces:
      - application/json
      responses:
        "200":
          description: Token
          schema:
            type: string
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      tags:
      - public
    put:
      consumes:
      - application/json
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/handlers.payload'
      produces:
      - application/json
      responses:
        "200":
          description: Token
          schema:
            type: string
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      tags:
      - public
  /rate:
    get:
      consumes:
      - application/json
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: integer
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User Rate
          schema:
            $ref: '#/definitions/models.UserRate'
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      tags:
      - rate
    post:
      consumes:
      - application/json
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.UserRateCreate'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully created
          schema:
            type: integer
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      tags:
      - rate
  /rate/claim:
    put:
      consumes:
      - application/json
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: amount
        in: query
        name: amount
        required: true
        type: integer
      - description: userRateID
        in: query
        name: user_rate_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      tags:
      - rate
  /rate/user:
    get:
      consumes:
      - application/json
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Page
        in: query
        name: page
        required: true
        type: integer
      - description: Per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Array
          schema:
            items:
              $ref: '#/definitions/models.UserRate'
            type: array
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      tags:
      - rate
  /user:
    get:
      consumes:
      - application/json
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      tags:
      - user
swagger: "2.0"